Struct aws_sdk_forecast::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonForecast.

This client allows ergonomic access to a AmazonForecast-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_dataset(&self) -> CreateDataset<C, M, R>

Constructs a fluent builder for the CreateDataset operation.

See CreateDataset for more information about the operation and its arguments.

pub fn create_dataset_group(&self) -> CreateDatasetGroup<C, M, R>

Constructs a fluent builder for the CreateDatasetGroup operation.

See CreateDatasetGroup for more information about the operation and its arguments.

pub fn create_dataset_import_job(&self) -> CreateDatasetImportJob<C, M, R>

Constructs a fluent builder for the CreateDatasetImportJob operation.

See CreateDatasetImportJob for more information about the operation and its arguments.

pub fn create_forecast(&self) -> CreateForecast<C, M, R>

Constructs a fluent builder for the CreateForecast operation.

See CreateForecast for more information about the operation and its arguments.

pub fn create_forecast_export_job(&self) -> CreateForecastExportJob<C, M, R>

Constructs a fluent builder for the CreateForecastExportJob operation.

See CreateForecastExportJob for more information about the operation and its arguments.

pub fn create_predictor(&self) -> CreatePredictor<C, M, R>

Constructs a fluent builder for the CreatePredictor operation.

See CreatePredictor for more information about the operation and its arguments.

pub fn create_predictor_backtest_export_job(
&self
) -> CreatePredictorBacktestExportJob<C, M, R>

Constructs a fluent builder for the CreatePredictorBacktestExportJob operation.

See CreatePredictorBacktestExportJob for more information about the operation and its arguments.

pub fn delete_dataset(&self) -> DeleteDataset<C, M, R>

Constructs a fluent builder for the DeleteDataset operation.

See DeleteDataset for more information about the operation and its arguments.

pub fn delete_dataset_group(&self) -> DeleteDatasetGroup<C, M, R>

Constructs a fluent builder for the DeleteDatasetGroup operation.

See DeleteDatasetGroup for more information about the operation and its arguments.

pub fn delete_dataset_import_job(&self) -> DeleteDatasetImportJob<C, M, R>

Constructs a fluent builder for the DeleteDatasetImportJob operation.

See DeleteDatasetImportJob for more information about the operation and its arguments.

pub fn delete_forecast(&self) -> DeleteForecast<C, M, R>

Constructs a fluent builder for the DeleteForecast operation.

See DeleteForecast for more information about the operation and its arguments.

pub fn delete_forecast_export_job(&self) -> DeleteForecastExportJob<C, M, R>

Constructs a fluent builder for the DeleteForecastExportJob operation.

See DeleteForecastExportJob for more information about the operation and its arguments.

pub fn delete_predictor(&self) -> DeletePredictor<C, M, R>

Constructs a fluent builder for the DeletePredictor operation.

See DeletePredictor for more information about the operation and its arguments.

pub fn delete_predictor_backtest_export_job(
&self
) -> DeletePredictorBacktestExportJob<C, M, R>

Constructs a fluent builder for the DeletePredictorBacktestExportJob operation.

See DeletePredictorBacktestExportJob for more information about the operation and its arguments.

pub fn delete_resource_tree(&self) -> DeleteResourceTree<C, M, R>

Constructs a fluent builder for the DeleteResourceTree operation.

See DeleteResourceTree for more information about the operation and its arguments.

pub fn describe_dataset(&self) -> DescribeDataset<C, M, R>

Constructs a fluent builder for the DescribeDataset operation.

See DescribeDataset for more information about the operation and its arguments.

pub fn describe_dataset_group(&self) -> DescribeDatasetGroup<C, M, R>

Constructs a fluent builder for the DescribeDatasetGroup operation.

See DescribeDatasetGroup for more information about the operation and its arguments.

pub fn describe_dataset_import_job(&self) -> DescribeDatasetImportJob<C, M, R>

Constructs a fluent builder for the DescribeDatasetImportJob operation.

See DescribeDatasetImportJob for more information about the operation and its arguments.

pub fn describe_forecast(&self) -> DescribeForecast<C, M, R>

Constructs a fluent builder for the DescribeForecast operation.

See DescribeForecast for more information about the operation and its arguments.

pub fn describe_forecast_export_job(&self) -> DescribeForecastExportJob<C, M, R>

Constructs a fluent builder for the DescribeForecastExportJob operation.

See DescribeForecastExportJob for more information about the operation and its arguments.

pub fn describe_predictor(&self) -> DescribePredictor<C, M, R>

Constructs a fluent builder for the DescribePredictor operation.

See DescribePredictor for more information about the operation and its arguments.

pub fn describe_predictor_backtest_export_job(
&self
) -> DescribePredictorBacktestExportJob<C, M, R>

Constructs a fluent builder for the DescribePredictorBacktestExportJob operation.

See DescribePredictorBacktestExportJob for more information about the operation and its arguments.

pub fn get_accuracy_metrics(&self) -> GetAccuracyMetrics<C, M, R>

Constructs a fluent builder for the GetAccuracyMetrics operation.

See GetAccuracyMetrics for more information about the operation and its arguments.

pub fn list_dataset_groups(&self) -> ListDatasetGroups<C, M, R>

Constructs a fluent builder for the ListDatasetGroups operation.

See ListDatasetGroups for more information about the operation and its arguments.

pub fn list_dataset_import_jobs(&self) -> ListDatasetImportJobs<C, M, R>

Constructs a fluent builder for the ListDatasetImportJobs operation.

See ListDatasetImportJobs for more information about the operation and its arguments.

pub fn list_datasets(&self) -> ListDatasets<C, M, R>

Constructs a fluent builder for the ListDatasets operation.

See ListDatasets for more information about the operation and its arguments.

pub fn list_forecast_export_jobs(&self) -> ListForecastExportJobs<C, M, R>

Constructs a fluent builder for the ListForecastExportJobs operation.

See ListForecastExportJobs for more information about the operation and its arguments.

pub fn list_forecasts(&self) -> ListForecasts<C, M, R>

Constructs a fluent builder for the ListForecasts operation.

See ListForecasts for more information about the operation and its arguments.

pub fn list_predictor_backtest_export_jobs(
&self
) -> ListPredictorBacktestExportJobs<C, M, R>

Constructs a fluent builder for the ListPredictorBacktestExportJobs operation.

See ListPredictorBacktestExportJobs for more information about the operation and its arguments.

pub fn list_predictors(&self) -> ListPredictors<C, M, R>

Constructs a fluent builder for the ListPredictors operation.

See ListPredictors for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn stop_resource(&self) -> StopResource<C, M, R>

Constructs a fluent builder for the StopResource operation.

See StopResource for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_dataset_group(&self) -> UpdateDatasetGroup<C, M, R>

Constructs a fluent builder for the UpdateDatasetGroup operation.

See UpdateDatasetGroup for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.