Module aws_sdk_forecast::error

Modules

invalid_input_exception

See InvalidInputException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

Structs

CreateDatasetError

CreateDatasetGroupError

CreateDatasetImportJobError

CreateForecastError

CreateForecastExportJobError

CreatePredictorBacktestExportJobError

CreatePredictorError

DeleteDatasetError

DeleteDatasetGroupError

DeleteDatasetImportJobError

DeleteForecastError

DeleteForecastExportJobError

DeletePredictorBacktestExportJobError

DeletePredictorError

DeleteResourceTreeError

DescribeDatasetError

DescribeDatasetGroupError

DescribeDatasetImportJobError

DescribeForecastError

DescribeForecastExportJobError

DescribePredictorBacktestExportJobError

DescribePredictorError

GetAccuracyMetricsError

InvalidInputException

We can't process the request because it includes an invalid value or a value that exceeds

InvalidNextTokenException

The token is not valid. Tokens expire after 24 hours.

LimitExceededException

The limit on the number of resources per account has been exceeded.

ListDatasetGroupsError

ListDatasetImportJobsError

ListDatasetsError

ListForecastExportJobsError

ListForecastsError

ListPredictorBacktestExportJobsError

ListPredictorsError

ListTagsForResourceError

ResourceAlreadyExistsException

There is already a resource with this name. Try again with a different name.

ResourceInUseException

The specified resource is in use.

ResourceNotFoundException

We can't find a resource with that Amazon Resource Name (ARN). Check the ARN and try

StopResourceError

TagResourceError

UntagResourceError

UpdateDatasetGroupError

Enums

CreateDatasetErrorKind

CreateDatasetGroupErrorKind

CreateDatasetImportJobErrorKind

CreateForecastErrorKind

CreateForecastExportJobErrorKind

CreatePredictorBacktestExportJobErrorKind

CreatePredictorErrorKind

DeleteDatasetErrorKind

DeleteDatasetGroupErrorKind

DeleteDatasetImportJobErrorKind

DeleteForecastErrorKind

DeleteForecastExportJobErrorKind

DeletePredictorBacktestExportJobErrorKind

DeletePredictorErrorKind

DeleteResourceTreeErrorKind

DescribeDatasetErrorKind

DescribeDatasetGroupErrorKind

DescribeDatasetImportJobErrorKind

DescribeForecastErrorKind

DescribeForecastExportJobErrorKind

DescribePredictorBacktestExportJobErrorKind

DescribePredictorErrorKind

GetAccuracyMetricsErrorKind

ListDatasetGroupsErrorKind

ListDatasetImportJobsErrorKind

ListDatasetsErrorKind

ListForecastExportJobsErrorKind

ListForecastsErrorKind

ListPredictorBacktestExportJobsErrorKind

ListPredictorsErrorKind

ListTagsForResourceErrorKind

StopResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDatasetGroupErrorKind