Module aws_sdk_forecast::model

Modules

categorical_parameter_range

See CategoricalParameterRange

continuous_parameter_range

See ContinuousParameterRange

data_destination

See DataDestination

data_source

See DataSource

dataset_group_summary

See DatasetGroupSummary

dataset_import_job_summary

See DatasetImportJobSummary

dataset_summary

See DatasetSummary

encryption_config

See EncryptionConfig

error_metric

See ErrorMetric

evaluation_parameters

See EvaluationParameters

evaluation_result

See EvaluationResult

featurization

See Featurization

featurization_config

See FeaturizationConfig

featurization_method

See FeaturizationMethod

filter

See Filter

forecast_export_job_summary

See ForecastExportJobSummary

forecast_summary

See ForecastSummary

hyper_parameter_tuning_job_config

See HyperParameterTuningJobConfig

input_data_config

See InputDataConfig

integer_parameter_range

See IntegerParameterRange

metrics

See Metrics

parameter_ranges

See ParameterRanges

predictor_backtest_export_job_summary

See PredictorBacktestExportJobSummary

predictor_execution

See PredictorExecution

predictor_execution_details

See PredictorExecutionDetails

predictor_summary

See PredictorSummary

s3_config

See S3Config

schema

See Schema

schema_attribute

See SchemaAttribute

statistics

See Statistics

supplementary_feature

See SupplementaryFeature

tag

See Tag

test_window_summary

See TestWindowSummary

weighted_quantile_loss

See WeightedQuantileLoss

window_summary

See WindowSummary

Structs

CategoricalParameterRange

Specifies a categorical hyperparameter and it's range of tunable values.

ContinuousParameterRange

Specifies a continuous hyperparameter and it's range of tunable values.

DataDestination

The destination for an export job. Provide an S3 path, an AWS Identity and Access Management (IAM) role that allows Amazon Forecast

DataSource

The source of your training data, an AWS Identity and Access Management (IAM) role that allows Amazon Forecast to

DatasetGroupSummary

Provides a summary of the dataset group properties used in the ListDatasetGroups operation. To get the complete set of properties, call the

DatasetImportJobSummary

Provides a summary of the dataset import job properties used in the ListDatasetImportJobs operation. To get the complete set of properties, call the

DatasetSummary

Provides a summary of the dataset properties used in the ListDatasets

EncryptionConfig

An AWS Key Management Service (KMS) key and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to

ErrorMetric

Provides detailed error metrics to evaluate the performance of a predictor. This object

EvaluationParameters

Parameters that define how to split a dataset into training data and testing data, and the

EvaluationResult

The results of evaluating an algorithm. Returned as part of the GetAccuracyMetrics response.

Featurization

Provides featurization (transformation) information for a dataset field. This object

FeaturizationConfig

In a CreatePredictor operation, the specified algorithm

FeaturizationMethod

Provides information about the method that featurizes (transforms) a dataset field.

Filter

Describes a filter for choosing a subset of objects. Each filter consists of a

ForecastExportJobSummary

Provides a summary of the forecast export job properties used in the ListForecastExportJobs operation. To get the complete set of properties, call

ForecastSummary

Provides a summary of the forecast properties used in the ListForecasts

HyperParameterTuningJobConfig

Configuration information for a hyperparameter tuning job. You specify this object in

InputDataConfig

The data used to train a predictor. The data includes a dataset group and any

IntegerParameterRange

Specifies an integer hyperparameter and it's range of tunable values.

Metrics

Provides metrics that are used to evaluate the performance of a predictor. This object is

ParameterRanges

Specifies the categorical, continuous, and integer hyperparameters, and their ranges of

PredictorBacktestExportJobSummary

Provides a summary of the predictor backtest export job properties used in the ListPredictorBacktestExportJobs operation. To get a complete set of properties, call

PredictorExecution

The algorithm used to perform a backtest and the status of those tests.

PredictorExecutionDetails

Contains details on the backtests performed to evaluate the accuracy of the predictor. The

PredictorSummary

Provides a summary of the predictor properties that are used in the ListPredictors operation. To get the complete set of properties, call the DescribePredictor operation, and provide the listed

S3Config

The path to the file(s) in an Amazon Simple Storage Service (Amazon S3) bucket, and an AWS Identity and Access Management (IAM) role that

Schema

Defines the fields of a dataset. You specify this object in the CreateDataset request.

SchemaAttribute

An attribute of a schema, which defines a dataset field. A schema attribute is required

Statistics

Provides statistics for each data field imported into to an Amazon Forecast dataset with

SupplementaryFeature

Describes a supplementary feature of a dataset group. This object is part of the InputDataConfig object. Forecast supports the Weather Index and Holidays built-in

Tag

The optional metadata that you apply to a resource to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

TestWindowSummary

The status, start time, and end time of a backtest, as well as a failure reason if

WeightedQuantileLoss

The weighted loss value for a quantile. This object is part of the Metrics object.

WindowSummary

The metrics for a time range within the evaluation portion of a dataset. This object is

Enums

AttributeType

AutoMlOverrideStrategy

DatasetType

Domain

EvaluationType

FeaturizationMethodName

FilterConditionString

OptimizationMetric

ScalingType