Module aws_sdk_forecast::operation

Structs

CreateDataset

Creates an Amazon Forecast dataset. The information about the dataset that you provide helps

CreateDatasetGroup

Creates a dataset group, which holds a collection of related datasets. You can add

CreateDatasetImportJob

Imports your training data to an Amazon Forecast dataset. You provide the location of your

CreateForecast

Creates a forecast for each item in the TARGET_TIME_SERIES dataset that was

CreateForecastExportJob

Exports a forecast created by the CreateForecast operation to your

CreatePredictor

Creates an Amazon Forecast predictor.

CreatePredictorBacktestExportJob

Exports backtest forecasts and accuracy metrics generated by the CreatePredictor operation. Two folders containing CSV files are exported

DeleteDataset

Deletes an Amazon Forecast dataset that was created using the CreateDataset

DeleteDatasetGroup

Deletes a dataset group created using the CreateDatasetGroup operation.

DeleteDatasetImportJob

Deletes a dataset import job created using the CreateDatasetImportJob

DeleteForecast

Deletes a forecast created using the CreateForecast operation. You can

DeleteForecastExportJob

Deletes a forecast export job created using the CreateForecastExportJob

DeletePredictor

Deletes a predictor created using the CreatePredictor operation. You can

DeletePredictorBacktestExportJob

Deletes a predictor backtest export job.

DeleteResourceTree

Deletes an entire resource tree. This operation will delete the parent resource and

DescribeDataset

Describes an Amazon Forecast dataset created using the CreateDataset

DescribeDatasetGroup

Describes a dataset group created using the CreateDatasetGroup

DescribeDatasetImportJob

Describes a dataset import job created using the CreateDatasetImportJob

DescribeForecast

Describes a forecast created using the CreateForecast operation.

DescribeForecastExportJob

Describes a forecast export job created using the CreateForecastExportJob operation.

DescribePredictor

Describes a predictor created using the CreatePredictor

DescribePredictorBacktestExportJob

Describes a predictor backtest export job created using the CreatePredictorBacktestExportJob operation.

GetAccuracyMetrics

Provides metrics on the accuracy of the models that were trained by the CreatePredictor operation. Use metrics to see how well the model performed and

ListDatasetGroups

Returns a list of dataset groups created using the CreateDatasetGroup

ListDatasetImportJobs

Returns a list of dataset import jobs created using the CreateDatasetImportJob operation. For each import job, this operation returns a

ListDatasets

Returns a list of datasets created using the CreateDataset operation.

ListForecastExportJobs

Returns a list of forecast export jobs created using the CreateForecastExportJob operation. For each forecast export job, this operation

ListForecasts

Returns a list of forecasts created using the CreateForecast operation.

ListPredictorBacktestExportJobs

Returns a list of predictor backtest export jobs created using the CreatePredictorBacktestExportJob operation. This operation returns a summary

ListPredictors

Returns a list of predictors created using the CreatePredictor

ListTagsForResource

Lists the tags for an Amazon Forecast resource.

StopResource

Stops a resource.

TagResource

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are also deleted.

UntagResource

Deletes the specified tags from a resource.

UpdateDatasetGroup

Replaces the datasets in a dataset group with the specified datasets.